Validate the user input
I validated the user input by first standardizing all forms of capitalization of the game options using the title() function. Then used an if statement to continue with the game or quit if there is an invalid input. I will try to change the code to give the end user another chance after an invalid input.