Final Submission for the Deadline
All basic requirements for the "Rock, Paper, Scissors" Exercise are completed and finalized under the README.md file and the game.py file.

I have completed the first objective "customization" of the further exploration exercise. I have attempted to complete the last objective "testing of the same exercise.

GitBash runs all aspects of the code smoothly except for the last objective, "testing". The file is not properly being collected, and there is 1 error that occurs. Although the logic is written, I am not sure why I am having this issue.